ProjectManagement/src/test/java/com/ibm/Demo/PathValidatorImplTest.java
/*
 * package com.ibm.Demo;
 * 
 * import static org.junit.Assert.assertEquals; import static
 * org.junit.Assert.assertFalse; import static org.junit.Assert.assertTrue;
 * import static org.mockito.Mockito.mock; import static
 * org.mockito.Mockito.when;
 * 
 * import java.io.File; import java.io.IOException; import java.nio.file.Files;
 * import java.nio.file.InvalidPathException; import java.nio.file.Path; import
 * java.nio.file.Paths;
 * 
 * import org.junit.Before; import org.junit.Rule; import org.junit.Test; import
 * org.junit.jupiter.api.Disabled; import org.junit.rules.TemporaryFolder;
 * import org.springframework.beans.factory.annotation.Autowired;
 * 
 * import com.ibm.Demo.service.impl.PathValidatorImpl;
 * 
 * @Disabled public class PathValidatorImplTest {
 * 
 * @Autowired private PathValidatorImpl pathValidator;
 * 
 * @Rule public TemporaryFolder tempFolder = new TemporaryFolder();
 * 
 * @Before public void setUp() { pathValidator = new PathValidatorImpl(); }
 * 
 * @Test(expected = NullPointerException.class) public void
 * testValidateAndReturnAbsolutePath_NullPath() {
 * pathValidator.validateAndReturnAbsolutePath(null); }
 * 
 * @Test(expected = NullPointerException.class) public void
 * testValidateAndReturnAbsolutePath_EmptyPath() {
 * pathValidator.validateAndReturnAbsolutePath(""); }
 * 
 * @Test(expected = InvalidPathException.class) public void
 * testValidateAndReturnAbsolutePath_InvalidPath() {
 * pathValidator.validateAndReturnAbsolutePath("invalid:path"); }
 * 
 * @Test(expected = InvalidPathException.class) public void
 * testValidateAndReturnAbsolutePath_NonexistentPath() {
 * pathValidator.validateAndReturnAbsolutePath("nonexistent/path"); }
 * 
 * @Test public void testValidateAndReturnAbsolutePath_ValidPath() throws
 * IOException { // Create a temporary file inside the temporary folder File
 * tempFile = tempFolder.newFile("testFile.txt"); Path validPath =
 * tempFile.toPath();
 * 
 * // Validate the path using the temporary file's path Path result =
 * pathValidator.validateAndReturnAbsolutePath(validPath.toString());
 * 
 * // Assert the result assertEquals(validPath.toAbsolutePath(), result); }
 * 
 * @Test(expected = IOException.class) public void
 * 
 * testUnzip_NonexistentZipFile() throws IOException { Path zipFilePath =
 * Paths.get("nonexistent.zip"); pathValidator.unzip(zipFilePath); }
 * 
 * @Test public void testUnzip_ValidZipFile() throws IOException { // Create a
 * temporary directory for the unzipped files Path tempDir =
 * tempFolder.newFolder("unzipped").toPath();
 * 
 * // Mock the PathValidatorImpl to avoid actual file system operations
 * PathValidatorImpl pathValidatorMock = mock(PathValidatorImpl.class);
 * 
 * // Mock the behavior of the unzip method to return the temporary directory
 * path Path zipFilePath = Paths.get("src/test/resources/test.zip");
 * when(pathValidatorMock.unzip(zipFilePath)).thenReturn(tempDir);
 * 
 * // Test the unzip method with the mocked behavior Path unzippedPath =
 * pathValidatorMock.unzip(zipFilePath);
 * 
 * // Assert that the returned path is the same as the temporary directory path
 * assertEquals(tempDir, unzippedPath);
 * assertTrue(Files.isDirectory(unzippedPath)); }
 * 
 * @Test public void testIsMavenProject_ValidMavenProject() { // Mock the
 * PathValidatorImpl PathValidatorImpl pathValidatorMock =
 * mock(PathValidatorImpl.class);
 * 
 * // Define the behavior of isMavenProject to return true for a specific path
 * Path directoryPath = Paths.get("src/test/resources/mavenProject");
 * when(pathValidatorMock.isMavenProject(directoryPath)).thenReturn(true);
 * 
 * // Test the method with the mocked behavior
 * assertTrue(pathValidatorMock.isMavenProject(directoryPath)); }
 * 
 * @Test public void testIsMavenProject_InvalidMavenProject() { Path
 * directoryPath = Paths.get("src/test/resources/nonMavenProject");
 * assertFalse(pathValidator.isMavenProject(directoryPath)); } }
 */